Work around "fake" empty path by uri-common (#1448)
When we check to see if the underlying path must be updated
to be present, we should know that uri-common is faking a
path of (/ "") when the path is empty.

The path normalization in uri-common could use some work though,
as it never really touches the underlying uri-generic object.
On the other hand, this is sort of the promise that uri-common
tries to make - it never mangles the uri object unless you explicitly
update something yourself.